pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "text-to-sql-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-llms-google-genai>=0.6.2",
    "llama-index-workflows>=2.5.0",
    "pgsql>=2.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/text_to_sql_workflow"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.llamadeploy.workflows]
text-to-sql-workflow = "text_to_sql_workflow.main:workflow"

[tool.llamadeploy.ui]
directory = "ui"

[tool.llamadeploy]
name = "text-to-sql-workflow"
env_files = [".env", "ui/.env"]